refactored database configurations
