+++ backend/mms_app_backend/mms_app_backend/api/utils.py
@@ -1,4 +1,4 @@
-from ..database import SessionLocal
+from backend.mms_app_backend.mms_app_backend.configs.database_config import SessionLocal
